A lot & Groundwork finalized
We need to finish sorting algs, and read in functions. We need to
convert our .cpp files to .c files.  I copied them over, but I was
getting issues with the .c files, so I left the .cpp files in there. The
main file compiles, but when compiling grid or wordlist there are some
issues.