added matrix and except
I added the matrix and except header files. we get a warning because
"d_matrix.h" #includes "d_except.h" which contains the strstream
library. This library is out of date and therefore a warning message
appears telling us that it "may be removed without further notice at a
future date". Ignore this warning message for now. working on sorting
algs. We still need to make functs that reads a file and makes a matrix
for the grid class. Also, we need to finish the findMatches function.
I'm not sure what the binary search function needs to do, but that is
also something we need to do.